backend/requirements.txt
absl-py==1.1.0
asgiref==3.6.0
astunparse==1.6.3
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==39.0.0
cssselect==1.1.0
cycler==0.11.0
Django==4.1.5
django-cors-headers==3.13.0
django-expiring-token==1.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flatbuffers==1.12
fonttools==4.33.3
fpdf==1.7.2
gast==0.4.0
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
ijson==3.1.4
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
libclang==14.0.1
lxml==4.9.1
Markdown==3.3.7
matplotlib==3.5.2
mysql==0.0.3
mysql-connector-python==8.0.31
mysqlclient==2.1.1
numpy==1.23.0
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
typing_extensions==4.3.0
tzdata==2022.7
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
XlsxWriter==3.0.4
